+++ examples/app-crm/src/components/calendar/calendar/index.module.css
