+++ examples/app-crm/src/components/calendar/categories/index.module.css
