+++ examples/app-crm/src/components/calendar/index.ts
@@ -1,0 +1,4 @@
+export * from "./calendar";
+export * from "./upcoming-events";
+export * from "./categories";
+export * from "./calendar-cell";
