+++ examples/app-crm/src/components/calendar/manage-categories/index.tsx
