+++ examples/app-crm/src/components/calendar/upcoming-events/event/index.tsx
