+++ examples/app-crm/src/components/calendar/upcoming-events/index.tsx
